+++ packages/regex-wand/tsup.config.ts
@@ -1,10 +1,11 @@
 import { defineConfig } from "tsup"
 
 export default defineConfig({
-	entry: ["src/index.ts"],
+	entry: ["src/index.ts", "src/transform.ts"],
 	format: ["esm"],
 	dts: true,
 	sourcemap: true,
+	splitting: false,
 	clean: true,
 	external: ["arkregex", "magic-regexp"],
 })